requirements.txt
torch
torchvision
ultralytics
streamlit
opencv-python-headless
pillow
numpy
scikit-learn
matplotlib
seaborn